+++ mod.ts
@@ -17,6 +17,3 @@ export {
   JSONBigIntReviver,
 } from './src/utils.ts'
 export { parseArkiveManifest } from './src/arkiver/manifest-validator.ts'
-export { ArkiveLib, type SourceInfo } from './src/lib/ArkiveLib.ts'
-export { Erc721Lib, type Erc721Opts } from './src/lib/nft/Erc721Lib.ts'
-export { AaveLib, type AaveOpts } from './src/lib/aave/AaveLib.ts'