+++ src/utils.ts
@@ -1,4 +1,4 @@
-import { Chains } from './arkiver/manifest-builder/manifest.ts'
+import { Chains } from './arkiver/types.ts'
 import { Arkive, SafeLog, SafeRpcLog } from './arkiver/types.ts'
 import { supportedChains } from './chains.ts'
 
@@ -27,10 +27,10 @@ export function assertChain(
 }
 
 export const bigIntMax = (...args: bigint[]) =>
-  args.reduce((m, e) => e > m ? e : m)
+  args.reduce((m, e) => e > m ? e : m, 0n)
 
 export const bigIntMin = (...args: bigint[]) =>
-  args.reduce((m, e) => e < m ? e : m)
+  args.reduce((m, e) => e < m ? e : m, 2n ** 256n)
 
 export function formatLog(
   log: SafeRpcLog,